requirements.txt
annoy
openai
python-dotenv
matplotlib
anthropic